+++ GPM550X/AppVersion.cs
@@ -2,7 +2,7 @@ namespace Gpm
 {
     internal static class AppVersion
     {
-        public const string Version = "1.0.0";
+        public const string Version = "1.0.1";
         public const string Guid = "com.mursisru.gpm550x";
         public const string Name = "GPM-550X";
     }